LPS-104618 - Remove Permissions component
Also move TopperBar into it's own file
